Fix for BigDecimal generator failure.
The act of constructing a BigDecimal with limited precision will cause
it to round, which can shift the scale and push it out of bounds of
what BigDecimal is willing to handle.  This causes an ArithmeticException
to be thrown.

The fix is simply to limit the scale such that the rounding process won't
go that far.