Readded Test.Params.maxDiscardedTests as deprecated.
Hopefully, ScalaCheck 1.10 should now be API-compatible with 1.9.
If maxDiscardedTests is used, it will override maxDiscardRatio.
(A new value will be assigned to maxDiscardRatio, which will be
maxDiscardedTests/minSuccessfulTests.)

This deprecation causes warnings when compiling ScalaCheck, but
we can't do anything about it, other than wait for 1.11 when
it can be removed.