Inline scala-parser-combinators
The cost of having Scalacheck depened on something other than
the standard libary is higher than the price or these 1000 lines
of copy/paste.

Such is the price of being a dependency of scala/scala.

Otherwise, project upstream that depend on scalacheck_2.11.0-M6
get scala-parser-combinators_2.11.0-M5 on their classpath, as
witnessed today in Scalaz [1] and earlier within the Scala modules
themselves.

> Okay, the build failed for scalacheck-binding:
>
> [error] Modules were resolved with conflicting cross-version suffixes in {file:/home/travis/build/scalaz/scalaz/}scalacheck-binding:
> [error]    org.scala-lang.modules:scala-xml _2.11.0-M6, _2.11.0-M5
> [error]    org.scala-lang.modules:scala-parser-combinators _2.11.0-M6, _2.11.0-M5
> java.lang.RuntimeException: Conflicting cross-version suffixes in: org.scala-lang.modules:scala-xml, org.scala-lang.modules:scala-parser-combinators

[1] https://github.com/scalaz/scalaz/pull/568#issuecomment-26674916