Add an optional name prefix to Properties.include().
This is useful when Properties with generic property names are included,
but are better described by the "parent" Properties object. That is,
include() without a prefix is a pure import facility, but with a prefix
it models a more hierarchical property structure.

This would replace a common helper function seen in various projects
like scalaz:

```
def checkAll(name: String, props: Properties) =
  for ((name2, prop) <- props.properties) yield {
    property(name + ":" + name2) = prop
  }
```