Support ScalaJS 0.6.0 (fix #122)
This commit is heavily inspired by 6e369b8 from @InTheKnow, but
much simplified. This commit does not try to merge the SBT framework
for JVM and JS, instead it only focuses implementing support for
ScalaJS 0.6.0. We should try to merge as much of the JVM/JS code
as possible in the future, though.