Initial work to move away from scala.util.Random.
This commit introduces a (seed: Long) parameter to many methods
that need to generate random values, as well as an Rng object
that implements a simple linear-congruent RNG.

The jvm tests pass. Currently getting permgen/OOM errors in the
js tests, not sure if this indicates a problem with the
implementation.

Still to do: there are still probably places where seeds are
not being threaded through correctly.