Add basic Cogen type.
This type allows values to perturb the RNG, meaning that we can
build something resembling a true Gen[A => B]. The basic strategy
is as follows:

Given a Cogen[A] and a Gen[B], we can create an A => B. When we
are given a particular A value, we use Cogen[A] to permute the
RNG seed we have, and then use Gen[B] to generate a B value.
In this way, the B values created are dependent on the A values
provided, and we have a "real" function.

This commit doesn't actually do any of that yet, but it does
introduce a Cogen[A] type, and define a bunch of useful instances.