Add .function1 and .function2 to create "true" function generators.
There is some ugliness in here (for example, .retrieve.get is a
pattern I've copied from elsewhere in the code). However, the
instances I've produced do seem to work.

The next step is to figure out how to use code-gen to create these
methods, as well as updating the Arbitrary code.