Update the Arbitrary type class instances to use Cogen.
This commit updates the code-generation script to use the
Gen.functionN methods to produce better arbitrary functions.

This is a breaking API change. We probably need to add more
Cogen instances for end-users to avoid breaking their tests.
However, Scalacheck's own tests still pass.

At this point the feature is added, and what is needed is
clean up and fleshing out the stable of Cogen instances.