Improve shrinking for integral/fractional types.
The previous code has to jump through a lot of hoops to try to support
both fractional and integral code with one class (based on
Numeric). Splitting the problem into two classes (and using the
appropriate division/quotient operation for each) makes it a lot
easier to see what is going on in each case.