Add Gen.pureApply with retry logic.
This commit adds a new method to Gen: pureApply. The idea is that this
method repeatedly calls doApply until some value is found, or until
the number of retries allowed is exhausted.

This functionality is especially useful in the function generators,
although it's possible to imagine third-party code that would take
advantage of this method as well.