Throw errors on invalid generators.
When generating generators whose parameters make valid values
impossible (e.g. choose(10, 0)), throw an exception rather than
returning an empty generator.

This commit also improves the logic of the Choose instances for
numbers, to make them able to generate values more efficiently when
generating the "entire range" of a number type.

Several of the tests have to be changed after this, to check for
exceptions when impossible parameter values are used.