Support deterministic testing using Parameters.
This commit adds an `initialSeed` parameter to the Parameter types
that Test and Gen use to control test execution. When set to `None`,
things behave as they usually do, but when set to `Some(seed)`, that
seed will be used to start the execution.

This makes it easier to configure repeatable tests. In the future, you
could imagine ScalaCheck emitting which seed it started with during a
failing test, and providing an interface for re-running the failing
test with the same seed.