Make Cogen invariant.
With the addition of Cogen instances for arbitrary function arities we were beginning to have implicit resolution issues with Cogen. For example, Cogen[Int => Int] was not able to find an implicit. This is related to a known issue with implicit resolution of contravariant types (see SI-2509). This is resolved by making Cogen invariant. Cogen instances can still be provided for a type and its subtypes using upper type bounds. See, for example, cogenSeq.