Fix non-terminating shrink streams.
This was reported as #244. The root cause was that shrinking certain
values would produce an infinite stream. There were several ways this
could happen -- in the reported case, the fact that Char was unsigned
meant that the strategy of including inverses was able to produce
loops.

The fix was to detect when Integral[T] cannot support negative values,
and disabling negation.

While working on this, I noticed that shrinking floating point
sentinels (i.e. infinities and NaN) would cause a similar issue. I
added code to detect those cases as well.

Fixes #244.