Fix issue where return code could be seen as 0.
Previously ScalaCheck tried to encode (and return) the number of
failing properties in the exit code. Unfortunately this meant that
multiples of 256 would be seen as having 0 failures.

The new code uses -1 for invalid arguments, 0 for success, and 1 for
one or more failures. The number of failing properties is also printed
explicitly, so that the user can still see this information.

Fixes #240.