genInfiniteStream
Adds a Gen instance for generating an infinite stream of values given a generator for the type of the underlying values. This is useful for testing that functions designed to be able to handle infinite data structures (e.g. through lazy evaluation) are working correctly. Also includes a new test for this instance.