Improve retryUntil's behavior during failure.
Previously retryUntil would throw a StackOverflowException if the
generator failed too many times. It's not obvious to users that the
problem is their generator and/or predicate.

This commit replaces this implicit depth limit with an explicit cutoff
after 10000 tries (along with a custom Exception). It also provides a
two-argument variant of retryUntil that allows the user to specify how
many tries are allowed.

This change should improve the situation in #232.