Improve BigInt and BigDecimal generators.
Previously, the BigDecimal generator could create astronomically-large
numbers that were not safe to use in most tests (so large that working
with them would cause seemingly-infinite hangs). This was primarily
due to using incredibly large scales (e.g. 10^2000000000). In practice
most users of BigDecimal are probably not trying to work with numbers
that large.

This commit also tries to even out the overall distributions of both
BigInt and BigDecimal numbers, while ensuring that none of the values
are so large that they will break most interesting properties. It's
relatively easy for users to write custom generators if they want
extremely large numbers.

Related to #246.