Reduce the frequency of None in Option Gen
The idea here is that if you are generating 100 values, it's probably
not that useful for 50 of them to be the same (None). Having a 50%
chance of None is especially problematic when a data structure has
several nested Option fields, in which case it becomes pretty unlikely
that any non-None values are generated for the innermost Option field.

I've arbitrarily chosen a 10% None frequency here, but feel free to
recommend other values that may be better.