Make Gen[Calendar] deterministic.
Previously it was possible for the "current system time" to leak
through into generated Calendar instance (and by extention Date
instances). This was especially painful for generated functions (such
as A => Date) where the function is likely to be evaluated at multiple
different system times.

This PR restructures Gen.calendar to ensure that all instances are
initialized from a generated time (in milliseconds). The code is
cleaned up a bit and a test was added to ensure generation is
deterministic.

Fixes #295.