Improve the seed interface a bit.
The major wart this fixes is the return value of fromBase64. This
method now returns Try[Seed], since it is expected to deal with
arbitrary user input.

This commit also improves the string representation of seeds, and
makes the seed internals public (which should be fine, since the seed
is immutable).

Finally this commit includes a few assorted fixes.