Thought better about making a, b, c, d public.
It might be the case that doing this would paint us into a corner
regarding the RNG used. Also, if we are really going to make those
public, we should probably make Seed a case class as well (there
wouldn't be much point in the current interface if we're completely
committed to the internals).