Fix distribution of Gen.pick
The update of Gen.pick in a2100f68 should have selected elements with
a uniform distribution, but used the wrong value for modulus when
generating the next random number.  The modulus should be the count,
and not the number of values to pick.

The code looks like the "reservoir sampling" technique, but just needs
to generate a random number within the range of the current count of
the collection, so that elements of the beginning of the list get
greater weight than the end.