Update to Scala Native 0.3.1, fix tests
A test was deliberately provoking a NullPointerException to make sure
that exceptions are correctly caught. However, in Scala Native, null
dereferencing is undefined behavior, and such an exception cannot be
caught at the moment. The test is changed to call next on an empty
Iterator, as these exceptions can be caught with Scala Native.