Improve doc for Gen.pick
Gen.pick, Gen.someOf, and Gen.atLeastOne will give a randomly chosen
subset of elements, but not in a random order.  This is because
Gen.pick is implemented with reservoir sampling but without a shuffle
operation.  Adding shuffling might be possible, but it could impact
performance and would violate the "do one thing well" principle of
Gen.pick.

For now, just document the situation that results aren't permuted in
the User Guide and in the API docs.