Deprecate internal forAllShrink
Shrinking failed tests will be deprecated when an explicit
generator is given to forAll.

Going forward, there should be another way for users to
shrink failing tests.  Naming it forAllShrink is the obvious
candidate.  This old version of forAllShrink taking a should
be made private, and should also be updated to take an
explicit Shrink[T] rather than an explicit T => Stream[T].