Allow Properties.overrideParameters to work correctly.
Currently when run from SBT (i.e. ScalaCheckFramework),
the overrideParameters method on Properties doesn't
successfully override any parameters. This was because
of the way the framework handles overrides coming from
build configuration -- the framework unconditionally
writes all parameter values, even when not explicitly
set in the build.

This commit reverses the order that these overrides are
applied, so that build overrides (which are coarser) can
be overridden by overrides in specific properties.