Maintain property order for tests.
Previously we were iterating over a map, resulting in a different order than
the sequence used to build the map. In the relevant case, we can just loop
over the sequence instead.

(The map in question _is_ useful to avoid repeatedly scanning the sequence when
running certain properties by name.)