Workaround bad implicit conversion
`prettyAny` defines an implicit conversion from `Any` to `Pretty`, this
is really dangerous in general. Here the problem is that `Pretty` has a
`map` member, so Dotty ends up choosing this conversion when we call
`map` on an array. We work around this by first converting the array to
a list, but the implicit conversions in this file should really be
rethought.