Stack-safe implementation of Pretty.break.
The motivation for this is #476. I wanted to ensure we had stronger
tests for Pretty.break before refactoring it, and also thought that
this implementation was a little bit nicer.

This PR includes a test for stack-safety, a test to ensure we never
create lines longer than the maximum allowed, and that we can
reassemble the original string from the output. The old code passed
all of these (apart from the stack safety test which it failed), which
helps give confidence and that the new implementation is equivalent.

It also includes a Scaladoc method to explain what the method does.

Thanks to @GytisZ for reporting the issue and providing an initial PR.