Validate ScalaJS and Scala Native versions in plugins.sbt.
Since we pass explicit versions in release.sh, it's important to make sure it
doesn't get out of sync with our definitions in plugins.sbt. To that end, this
commit ensures that we get one of two exact versions we're prepared to support.
Otherwise, we'll throw an error during SBT initialization.

This makes it slightly more annoying to test ScalaCheck with a new version of
one of these plugins (you have to modify plugins.sbt as opposed to just passing
a new variable on the command line) but also means we'll never accidentally
release with the wrong version.

I ended up printing before throwing an error because SBT doesn't display the
actual exception (and message) unless you type (l)ast. This way, you'll always
see the message (which should make it pretty obvious what the mistake was).