Fix a bug with initial seed support.
Prevously, the initial seed was used for every single property
evaluation, meaning that even if a property is evaluated 100 times the
same values will be used each time. This is not what users expect, and
makes the library harder to use.

(In particular, the MUnit library hit this bug using Test.check.)

After this commit, the initial seed will be used first, but then a
deterministic set of seeds will be used later. This sequence will be
the same for every run as long as the same number of workers are used.
If the number of workers is changed, the initial seed will be the
same, but the exact sequence of seeds used may differ somewhat.