Optimize shrinking of integral types
So long as the test case keeps succeeding, shrink to the next query
value of a binary search that searches for the smallest failing value.

That is, try the midpoint between the most recent test failure and the
previous shrunk value: x => x * (1 - 1/2^k) for natural numbers k.

For example, `shrink(32).toList == List(0, 16, 24, 28, 30, 31)`.