Use Special Gen To Avoid Overflow
The tests for the off by one error rely upon adding 1 second to the generated value. This _can_ cause an overflow if the generated Instant/Duration is less than 1 second from the max value the type can support.

This commit adds special generators to the GenSpecification for java.time values to ensure this never occurs. No one likes a flaky test.