Verify only props matching the given TestSelectors (#1031)
* Verify only props matching the given TestSelectors

Previously, ScalaCheck ignored the selectors that it receives as input
in the "root task". This prevented users from running only a subset of
properties in a specification by passing their `TestSelector` to
ScalaCheck in a `TaskDef`.

This patch fixes this by having the root task program the execution of
only the requested properties if the `TaskDef` contains only
`TestSelector`s. ScalaCheck's behavior remains unchanged if the
`TaskDef` contains any other kind of `Selector`.

* Combine properties in test

Previously, only the last property would be verified and the others
would be ignored. This patch fixes the tests so that all properties are
verified.